+++ src/exports/utils.ts
@@ -2,3 +2,7 @@
 export { convertToEmailSafeHtml, EMAIL_SAFE_CONFIG } from '../utils/emailSafeHtml'
 export { validateEmailHtml } from '../utils/validateEmailHtml'
 export type { ValidationResult } from '../utils/validateEmailHtml'
+
+// Configuration utilities
+export { getBroadcastConfig } from '../utils/getBroadcastConfig'
+export { getResendConfig } from '../utils/getResendConfig'